feat: 분석 페이지 UI 개발
