+++ src/components/footer/index.tsx
@@ -1,7 +1,7 @@
 const Footer = () => {
   return (
-    <div className="flex w-full mt-auto pt-60 px-6 pb-6 justify-center items-center shrink-0 border-t border-[#dad9d9] bg-white mx-auto">
-      <div className="flex flex-col w-full h-42 shrink-0 text-[#929292] typo-title gap-2">
+    <div className="flex w-full mt-20 pt-5 px-6 pb-6 justify-center items-center shrink-0 border-t border-[#dad9d9] bg-white mx-auto">
+      <div className="flex flex-col w-full h-42 shrink-0 text-[#929292] typo-body1-medium gap-2">
         <div>CRiT</div>
         <div>Creator in Trend</div>
         <br />
