+++ src/components/header/index.tsx
@@ -14,7 +14,7 @@ const Header = () => {
       <CritLogo className="w-38 h-11" />
       <div className="flex justify-center items-center gap-5 self-stretch">
         <div
-          className="flex w-24 h-9 flex-col justify-center text-center cursor-pointer typo-body"
+          className="flex w-24 h-9 flex-col justify-center text-center cursor-pointer typo-body2"
           onClick={() => navigate('/recommend')}
         >
           <div className={currentPath === '/recommend' ? 'text-[#6B4EFF]' : 'text-black'}>
@@ -22,18 +22,18 @@ const Header = () => {
           </div>
         </div>
         <div
-          className="flex w-24 h-9 flex-col justify-center text-center cursor-pointer typo-body"
+          className="flex w-24 h-9 flex-col justify-center text-center cursor-pointer typo-body2"
           onClick={() => navigate('/analysis')}
         >
           <div className={currentPath === '/analysis' ? 'text-[#6B4EFF]' : 'text-black'}>
             채널분석
           </div>
         </div>
-        <div className="flex w-24 h-9 flex-col justify-center text-center cursor-not-allowed typo-body opacity-40">
+        <div className="flex w-24 h-9 flex-col justify-center text-center cursor-not-allowed typo-body2 opacity-40">
           <div className="text-black">트렌드</div>
         </div>
       </div>
-      <div className="flex w-40 h-9 justify-center items-center text-black typo-body">
+      <div className="flex w-40 h-9 justify-center items-center text-black typo-body2">
         name님 어서오세요!
       </div>
       <div className="flex h-11 items-center justify-end gap-5">
