+++ src/index.css
@@ -41,17 +41,28 @@
 }
 
 @layer utilities {
-  /* Typography - Header */
-  .typo-header {
-    font-size: 22px;
+  /* ===== 피그마 타이포 시스템 ===== */
+
+  /* typo-title1: 25px / semi bold / -2.5% / 140% */
+  .typo-title1 {
+    font-size: 25px;
     font-style: normal;
     font-weight: 600;
-    line-height: 20px;
-    letter-spacing: 0;
+    line-height: 140%;
+    letter-spacing: -0.625px;
   }
 
-  /* Typography - Title */
-  .typo-title {
+  /* typo-title2: 30px / semi bold / -2.5% / 140% */
+  .typo-title2 {
+    font-size: 30px;
+    font-style: normal;
+    font-weight: 600;
+    line-height: 140%;
+    letter-spacing: -0.75px;
+  }
+
+  /* typo-body1-medium: 20px / medium / 0 / 20px */
+  .typo-body1-medium {
     font-size: 20px;
     font-style: normal;
     font-weight: 500;
@@ -59,7 +70,55 @@
     letter-spacing: 0;
   }
 
-  /* Typography - Title Bold */
+  /* typo-body2: 16px / regular / 0 / 20px */
+  .typo-body2 {
+    font-size: 16px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 20px;
+    letter-spacing: 0;
+  }
+
+  /* typo-body3: 16px / regular / 0 / 20px (Inter) */
+  .typo-body3 {
+    font-family: 'Inter', sans-serif;
+    font-size: 16px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 20px;
+    letter-spacing: 0;
+  }
+
+  /* typo-body4-semibold: 14px / semi bold / 0 / 20px */
+  .typo-body4-semibold {
+    font-size: 14px;
+    font-style: normal;
+    font-weight: 600;
+    line-height: 20px;
+    letter-spacing: 0;
+  }
+
+  /* typo-body5: 10px / regular / 0 / 10px */
+  .typo-body5 {
+    font-size: 10px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 20px;
+    letter-spacing: 0;
+  }
+
+  /* typo-body6: 10px / medium / 140% / 5px */
+  .typo-body6 {
+    font-size: 10px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: 140%;
+    letter-spacing: 5px;
+  }
+
+  /* ===== 피그마에 없지만 사용 중 ===== */
+
+  /* typo-title-bold: 20px / bold / 0 / 20px */
   .typo-title-bold {
     font-size: 20px;
     font-style: normal;
@@ -68,16 +127,7 @@
     letter-spacing: 0;
   }
 
-  /* Typography - Body */
-  .typo-body {
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 20px;
-    letter-spacing: 0;
-  }
-
-  /* Typography - Body Bold */
+  /* typo-body-bold: 16px / semi bold / 0 / 20px */
   .typo-body-bold {
     font-size: 16px;
     font-style: normal;
@@ -86,49 +136,13 @@
     letter-spacing: 0;
   }
 
-  /* Typography - Body */
+  /* typo-label: 12px / medium / 0 / 23px */
   .typo-label {
     font-size: 12px;
     font-style: normal;
     font-weight: 500;
     line-height: 23px;
     letter-spacing: 0;
-  }
-
-  /* Typography - h1 */
-  .typo-h1 {
-    font-size: 30px;
-    font-style: normal;
-    font-weight: 600;
-    line-height: 20px;
-    letter-spacing: 0;
-  }
-
-  /* Typography - Label2 */
-  .typo-label2 {
-    font-size: 14px;
-    font-style: normal;
-    font-weight: 600;
-    line-height: 20px;
-    letter-spacing: 0;
-  }
-
-  /* Typography - Title2 */
-  .typo-title2 {
-    font-size: 35px;
-    font-style: normal;
-    font-weight: 600;
-    line-height: 140%;
-    letter-spacing: -0.875px;
-  }
-
-  /* Typography - Body1 Medium */
-  .typo-body1-medium {
-    font-size: 20px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: 140%;
-    letter-spacing: 1px;
   }
 }
 
@@ -144,4 +158,20 @@
 .script-scroll::-webkit-scrollbar-thumb {
   background: #a594f9;
   border-radius: 1000px;
+}
+
+/* Fade in up animation */
+@keyframes fadeInUp {
+  from {
+    opacity: 0;
+    transform: translateY(30px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+.animate-fade-in-up {
+  animation: fadeInUp 0.6s ease-out both;
 }
