+++ src/pages/analysis/detailAnalysis/index.tsx
@@ -1,0 +1,9 @@
+const defailAnalysis = () => {
+  return (
+    <div>
+      <div> 분석 상세 페이지 개발 예정 </div>
+    </div>
+  );
+};
+
+export default defailAnalysis;
