+++ src/pages/recommend/index.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 import Header from '@/components/header';
-import TabList from '@/components/tabList';
-import FormList from '@/components/formList';
-import FormAnswer from '@/components/formAnswer';
-import FormSubject from '@/components/formSubject';
+import TabList from '@/components/recommendContent/tabList';
+import FormList from '@/components/recommendContent/formList';
+import FormAnswer from '@/components/recommendContent/formAnswer';
+import FormSubject from '@/components/recommendContent/formSubject';
 
 import Footer from '@/components/footer';
 
@@ -24,15 +24,19 @@ const RecommendPage = () => {
     <div>
       <Header />
       <div className="flex flex-col items-center px-40 gap-10">
-        <div className="flex flex-col items-center mt-20 w-full mx-auto">
+        <div className="flex flex-col items-center mt-20 w-full mx-auto animate-fade-in-up">
           <div className="relative z-10 mb-[-32px]">
             <TabList tabs={['롱폼', '숏폼']} />
           </div>
           <FormList onSearch={handleSearch} />
         </div>
-        {showSubject && <FormSubject onSelect={handleSelectSubject} />}
+        {showSubject && (
+          <div className="animate-fade-in-up">
+            <FormSubject onSelect={handleSelectSubject} />
+          </div>
+        )}
         {showAnswer && (
-          <div className="flex flex-col items-center self-stretch w-full mx-auto">
+          <div className="flex flex-col items-center self-stretch w-full mx-auto animate-fade-in-up">
             <FormAnswer />
           </div>
         )}
